Split off PreferencesWidget into separate module
