+++ gridsync/gui/widgets.py
@@ -3,16 +3,15 @@
 import json
 import os
 
-from PyQt5.QtCore import pyqtSignal, QPropertyAnimation, Qt, QThread
+from PyQt5.QtCore import QPropertyAnimation, QThread
 from PyQt5.QtGui import QColor, QPainter, QPixmap
 from PyQt5.QtWidgets import (
-    QCheckBox, QComboBox, QDialogButtonBox, QFileDialog, QFormLayout,
-    QGridLayout, QGroupBox, QLabel, QLineEdit, QPlainTextEdit, QProgressDialog,
-    QPushButton, QSizePolicy, QSpacerItem, QSpinBox, QWidget)
+    QComboBox, QDialogButtonBox, QFileDialog, QFormLayout, QGridLayout,
+    QGroupBox, QLabel, QLineEdit, QPlainTextEdit, QProgressDialog, QPushButton,
+    QSizePolicy, QSpacerItem, QSpinBox, QWidget)
 from gridsync.crypto import Crypter
 from gridsync.gui.password import PasswordDialog
 from gridsync.msg import error
-from gridsync.preferences import set_preference, get_preference
 
 
 class CompositePixmap(QPixmap):
@@ -282,71 +281,3 @@ class TahoeConfigForm(QWidget):
             error(self, type(e).__name__, str(e))
             return
         self.parse_content(content)
-
-
-class PreferencesWidget(QWidget):
-
-    accepted = pyqtSignal()
-
-    def __init__(self):
-        super(PreferencesWidget, self).__init__()
-        notifications_groupbox = QGroupBox("Notifications:", self)
-        notifications_label = QLabel("Show a desktop notification when...")
-        self.checkbox_connection = QCheckBox("Connection status changes")
-        self.checkbox_folder = QCheckBox("A folder is updated")
-        self.checkbox_invite = QCheckBox("An invite code is used")
-
-        notifications_layout = QGridLayout()
-        notifications_layout.addWidget(notifications_label)
-        notifications_layout.addWidget(self.checkbox_connection)
-        notifications_layout.addWidget(self.checkbox_folder)
-        notifications_layout.addWidget(self.checkbox_invite)
-        notifications_groupbox.setLayout(notifications_layout)
-        self.buttonbox = QDialogButtonBox(QDialogButtonBox.Ok)
-
-        layout = QGridLayout(self)
-        layout.addWidget(notifications_groupbox)
-        layout.addItem(QSpacerItem(0, 0, 0, QSizePolicy.Expanding))
-        layout.addWidget(self.buttonbox)
-
-        self.load_preferences()
-
-        self.checkbox_connection.stateChanged.connect(
-            self.on_checkbox_connection_changed)
-        self.checkbox_folder.stateChanged.connect(
-            self.on_checkbox_folder_changed)
-        self.checkbox_invite.stateChanged.connect(
-            self.on_checkbox_invite_changed)
-        self.buttonbox.accepted.connect(self.accepted.emit)
-
-    def load_preferences(self):
-        if get_preference('notifications', 'connection') == 'false':
-            self.checkbox_connection.setCheckState(Qt.Unchecked)
-        else:
-            self.checkbox_connection.setCheckState(Qt.Checked)
-        if get_preference('notifications', 'folder') == 'false':
-            self.checkbox_folder.setCheckState(Qt.Unchecked)
-        else:
-            self.checkbox_folder.setCheckState(Qt.Checked)
-        if get_preference('notifications', 'invite') == 'false':
-            self.checkbox_invite.setCheckState(Qt.Unchecked)
-        else:
-            self.checkbox_invite.setCheckState(Qt.Checked)
-
-    def on_checkbox_connection_changed(self, state):  # pylint:disable=no-self-use
-        if state:
-            set_preference('notifications', 'connection', 'true')
-        else:
-            set_preference('notifications', 'connection', 'false')
-
-    def on_checkbox_folder_changed(self, state):  # pylint:disable=no-self-use
-        if state:
-            set_preference('notifications', 'folder', 'true')
-        else:
-            set_preference('notifications', 'folder', 'false')
-
-    def on_checkbox_invite_changed(self, state):  # pylint:disable=no-self-use
-        if state:
-            set_preference('notifications', 'invite', 'true')
-        else:
-            set_preference('notifications', 'invite', 'false')
